requirements.txt
matplotlib == 3.5.2
ml_collections == 0.1.1
numpy
PyYAML == 6.0
tqdm==4.64.0
wfdb==3.4.1
fbm==0.3.0
evaluate
lib==4.0.0
pandas==1.3.5
requests==2.27.1
